+++ App/sample/commands/fun/ping.js
@@ -1,6 +1,8 @@
 const { SlashCommandBuilder } = require('discord.js');
 
 module.exports = {
+    category: 'fun',
+    cooldown: 60,
     data: new SlashCommandBuilder()
         .setName('ping')
         .setDescription('Replies with Pong!'),